+++ src/main/java/kiv/zcu/knowledgeipr/core/sourcedb/datasearch/mongo/MongoQueryRunner.java
@@ -40,16 +40,16 @@ public class MongoQueryRunner implements IQueryRunner {
      * {@inheritDoc}
      */
     @Override
-    public List<Pair<String, Long>> activePeople(DataSource collectionName, String type, int limit) {
+    public List<Pair<String, Integer>> activePeople(DataSource collectionName, String type, int limit) {
         LOGGER.info("Running 'active " + type + "' method on " + collectionName + " collection.");
 
-        List<Pair<String, Long>> activeAuthors = new ArrayList<>();
+        List<Pair<String, Integer>> activeAuthors = new ArrayList<>();
 
         AggregateIterable<Document> output = mongoRunner.runCountUnwindAggregation(collectionName, type, type + ".name", limit);
 
         for (Document doc : output) {
             String author = doc.get(type, Document.class).getString("name");
-            activeAuthors.add(new Pair<>(author, (Long) doc.get("count")));
+            activeAuthors.add(new Pair<>(author, (Integer) doc.get("count")));
         }
 
         return activeAuthors;
